+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### ✨ New features and improvements
 
+- Implement initial PythonTA LSP server
 - Add initial PythonTA implementation that simply calls the CLI
 
 ### 🐛 Bug fixes